+++ hoomd/dem/DEM3DForceComputeGPU.cc
@@ -153,7 +153,7 @@ void DEM3DForceComputeGPU<Real, Real4, Potential>::computeForces(uint64_t timest
     ArrayHandle<unsigned int> d_nlist(this->m_nlist->getNListArray(),
                                       access_location::device,
                                       access_mode::read);
-    ArrayHandle<unsigned int> d_head_list(this->m_nlist->getHeadList(),
+    ArrayHandle<size_t> d_head_list(this->m_nlist->getHeadList(),
                                           access_location::device,
                                           access_mode::read);
